+++ edx_toggles/tests/test_legacy_waffle.py
@@ -4,38 +4,28 @@ Unit tests for legacy waffle objects. These tests should be moved to test_waffle
 from django.test import TestCase
 
 from edx_toggles.toggles.internal.waffle.legacy import (
-    WaffleFlag,
-    WaffleFlagNamespace,
-    WaffleSwitch,
-    WaffleSwitchNamespace
+    LegacyWaffleFlag,
+    LegacyWaffleFlagNamespace,
+    LegacyWaffleSwitch,
+    LegacyWaffleSwitchNamespace
 )
 
 
-class TestWaffleSwitch(TestCase):
+class TestLegacyWaffleSwitch(TestCase):
     """
-    Tests the WaffleSwitch.
+    Tests the LegacyWaffleSwitch.
     """
-
-    def test_namespaced_switch_name(self):
-        """
-        Verify namespaced_switch_name returns the correct namespace switch name
-        """
-        namespace = WaffleSwitchNamespace("test_namespace")
-        switch = WaffleSwitch(namespace, "test_switch_name", __name__)
-        self.assertEqual(
-            "test_namespace.test_switch_name", switch.namespaced_switch_name
-        )
 
     def test_default_value(self):
-        namespace = WaffleSwitchNamespace("test_namespace")
-        switch = WaffleSwitch(namespace, "test_switch_name", module_name="module1")
+        namespace = LegacyWaffleSwitchNamespace("test_namespace")
+        switch = LegacyWaffleSwitch(namespace, "test_switch_name", module_name="module1")
         self.assertFalse(switch.is_enabled())
         self.assertFalse(namespace.is_enabled("test_switch_name"))
 
     # pylint: disable=protected-access
     def test_set_request_cache_with_short_name(self):
-        namespace = WaffleSwitchNamespace("test_namespace")
-        switch = WaffleSwitch(namespace, "test_switch_name", module_name="module1")
+        namespace = LegacyWaffleSwitchNamespace("test_namespace")
+        switch = LegacyWaffleSwitch(namespace, "test_switch_name", module_name="module1")
         self.assertFalse(switch._cached_switches.get("test_namespace.test_switch_name"))
         namespace.set_request_cache_with_short_name("test_switch_name", True)
         self.assertTrue(switch._cached_switches.get("test_namespace.test_switch_name"))
@@ -45,18 +35,13 @@ class TestWaffleSwitch(TestCase):
         self.assertFalse(switch.is_enabled())
 
 
-class TestWaffleFlag(TestCase):
+class TestLegacyWaffleFlag(TestCase):
     """
     Legacy waffle flag tests.
     """
 
-    def test_namespaced_flag_name(self):
-        namespace = WaffleFlagNamespace("namespace")
-        flag = WaffleFlag(namespace, "flag")
-        self.assertEqual("namespace.flag", flag.namespaced_flag_name)
-
     def test_default_value(self):
-        namespace = WaffleFlagNamespace("namespace")
-        flag = WaffleFlag(namespace, "flag")
+        namespace = LegacyWaffleFlagNamespace("namespace")
+        flag = LegacyWaffleFlag(namespace, "flag")
         self.assertFalse(flag.is_enabled())
         self.assertFalse(namespace.is_flag_active("flag"))